+++ app/models/test_file.rb
@@ -1,6 +1,5 @@
 class TestFile < ActiveRecord::Base
   belongs_to :test_list
-  has_many :strokes
 
   validates :name, presence: true
 end